requirements.txt
fastapi
uvicorn
pydantic
typer[all]
pyyaml
